+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 6.1.1 under development
 
-- no changes in this release.
+- Bug #112: Fix events configuration for Yii Debug (@vjik)
 
 ## 6.1.0 May 28, 2024
 
@@ -10,7 +10,7 @@
 - Enh #107: Implement lazy loading for injections (@vjik)
 - Enh #79: Add debug collector for yiisoft/yii-debug (@xepozz)
 - Enh #99: Make `viewPath` in `ViewRenderer` constructor optional (@vjik)
-- Bug #82: Fixed find for layout file due to compatibility with `yiisoft/view` (@rustamwin)
+- Bug #82: Fix find for layout file due to compatibility with `yiisoft/view` (@rustamwin)
 
 ## 6.0.0 February 16, 2023